# lib/api-zod/src/generated/types/classSession.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification for a campus app that shows which students are in class right now.
 * OpenAPI spec version: 0.1.0
 */

export interface ClassSession {
  courseCode: string;
  courseName: string;
  section: string;
  date: Date;
  /** 24-hour HH:MM in Asia/Kolkata time */
  startTime: string;
  /** 24-hour HH:MM in Asia/Kolkata time */
  endTime: string;
  /** @nullable */
  room: string | null;
}