# lib/api-zod/src/generated/types/scheduleDay.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification for a campus app that shows which students are in class right now.
 * OpenAPI spec version: 0.1.0
 */
import type { ClassSession } from './classSession';

export interface ScheduleDay {
  date: Date;
  sessions: ClassSession[];
}