# lib/api-zod/src/generated/types/studentStatus.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification for a campus app that shows which students are in class right now.
 * OpenAPI spec version: 0.1.0
 */
import type { ClassSession } from './classSession';
import type { StudentSummary } from './studentSummary';

export interface StudentStatus {
  student: StudentSummary;
  isInClass: boolean;
  currentSession: ClassSession | null;
  /** The next session later today, if any */
  nextSession: ClassSession | null;
}