# lib/api-zod/src/generated/types/studentSummary.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification for a campus app that shows which students are in class right now.
 * OpenAPI spec version: 0.1.0
 */
import type { Campus } from './campus';

export interface StudentSummary {
  id: number;
  name: string;
  email: string;
  campus: Campus;
  section: string;
}